Minor code enhancements

+++ prisma/migrations/20241030135444_init/migration.sql
@@ -3,14 +3,14 @@ CREATE TABLE "transactions" (
     "id" SERIAL NOT NULL,
     "timestamp" TIMESTAMPTZ NOT NULL,
     "status" BOOLEAN NOT NULL DEFAULT false,
-    "block_number" BIGINT NOT NULL,
+    "block_number" DECIMAL NOT NULL,
     "tx_index" INTEGER NOT NULL,
     "from" TEXT NOT NULL,
     "to" TEXT NOT NULL,
-    "value" BIGINT NOT NULL,
-    "gas_limit" BIGINT NOT NULL,
-    "gas_used" BIGINT NOT NULL,
-    "gas_price" BIGINT NOT NULL,
+    "value" DECIMAL NOT NULL,
+    "gas_limit" DECIMAL NOT NULL,
+    "gas_used" DECIMAL NOT NULL,
+    "gas_price" DECIMAL NOT NULL,
 
     CONSTRAINT "transactions_pkey" PRIMARY KEY ("id")
 );
